# gmc_api_gateway/docs/swagger.yaml
basePath: /gmcapi/v2
definitions:
...
  model.Active:
    properties:
      kind:
        type: string
      name:
        type: string
      namespace:
        type: string
    type: object
...
  model.CRONJOB:
    properties:
      active:
        items:
          $ref: '#/definitions/model.Active'
        type: array
      annotations: {}
      cluster:
        type: string
      concurrencyPolicy:
        type: string
      containers:
        items:
          $ref: '#/definitions/model.Containers'
        type: array
      creationTimestamp:
        type: string
      events:
        items:
          $ref: '#/definitions/model.EVENT'
        type: array
      failedJobsHistoryLimit:
        type: integer
      label: {}
      lastScheduleTime:
        type: string
      name:
        type: string
      project:
        type: string
      schedule:
        type: string
      successfulJobsHistoryLimit:
        type: integer
      user:
        type: string
      workspace:
        type: string
    type: object
...
  model.Containers:
    properties:
      image:
        type: string
      name:
        type: string
    type: object
...
  model.DAEMONSET_DETAIL:
    properties:
      NodeSelector: {}
      annotations: {}
      cluster:
        type: string
      containers: {}
      createAt:
        type: string
      events:
        items:
          $ref: '#/definitions/model.EVENT'
        type: array
      label: {}
      name:
        type: string
      project:
        type: string
      ready:
        type: string
      status: {}
      strategy: {}
      user:
        type: string
      workspace:
        type: string
    type: object
...
  model.EVENT:
    properties:
      cluster:
        type: string
      eventTime:
        type: string
      kind:
        type: string
      message:
        type: string
      name:
        type: string
      namespace:
        type: string
      reason:
        type: string
      type:
        type: string
    type: object
  model.Error:
    properties:
      data:
        description: Email    string `json:"email"`
        type: string
      status:
        type: string
    type: object
...
  model.WORKLOAD:
    properties:
      NodeSelector: {}
      cluster:
        type: string
      createAt:
        type: string
      name:
        type: string
      project:
        type: string
      ready:
        type: string
      user:
        type: string
      workspace:
        type: string
    type: object
...
host: 101.79.1.138:8012
info:
  contact:
    name: GM-Center
    url: https://gedge-platform.github.io/gm-center/
  description: This is a Gedge GM-Center Swagger API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gedge GM-Center Swagger API
  version: "2.0"
paths:
...
  /cronjobs:
    get:
      consumes:
      - application/json
      description: get Cronjob List
      parameters:
      - description: name of the Workspace
        in: query
        name: workspace
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        type: string
      - description: name of the Project
        in: query
        name: project
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CRONJOB'
      security:
      - Bearer: []
      summary: Show List Cronjob
      tags:
      - Kubernetes
    post:
      consumes:
      - application/json
      description: Create Cronjob
      parameters:
      - description: Cronjob Info Body
        in: body
        name: json
        required: true
        schema:
          type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CRONJOB'
      security:
      - Bearer: []
      summary: Create Cronjob
      tags:
      - Kubernetes
  /cronjobs/{name}:
    delete:
      consumes:
      - application/json
      description: Delete Cronjob
      parameters:
      - description: name of the Cronjob
        in: path
        name: name
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - Bearer: []
      summary: Delete Cronjob
      tags:
      - Kubernetes
    get:
      consumes:
      - application/json
      description: get Cronjob Details
      parameters:
      - description: name of the Cronjob
        in: path
        name: name
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CRONJOB'
      security:
      - Bearer: []
      summary: Show detail Cronjob
      tags:
      - Kubernetes
  /daemonsets:
    get:
      consumes:
      - application/json
      description: get Daemonset List
      parameters:
      - description: name of the Workspace
        in: query
        name: workspace
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        type: string
      - description: name of the Project
        in: query
        name: project
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WORKLOAD'
      security:
      - Bearer: []
      summary: Show List Daemonset
      tags:
      - Kubernetes
    post:
      consumes:
      - application/json
      description: Create Daemonset
      parameters:
      - description: Daemonset Info Body
        in: body
        name: json
        required: true
        schema:
          type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WORKLOAD'
      security:
      - Bearer: []
      summary: Create Daemonset
      tags:
      - Kubernetes
  /daemonsets/{name}:
    delete:
      consumes:
      - application/json
      description: Delete Daemonset
      parameters:
      - description: name of the Daemonset
        in: path
        name: name
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - Bearer: []
      summary: Delete Daemonset
      tags:
      - Kubernetes
    get:
      consumes:
      - application/json
      description: get Daemonset Details
      parameters:
      - description: name of the Daemonset
        in: path
        name: name
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      - description: name of the Project
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DAEMONSET_DETAIL'
      security:
      - Bearer: []
      summary: Show detail Daemonset
      tags:
      - Kubernetes
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: '"Type \"Bearer \" and then your API Token"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"